- Dév/évo/tech : A la fin des imports (Calc ou XML),on avait une erreur "en dehors du thread JavaFX" depuis qu'on a ajouét des listener pour màj la barre d'état des tâches (nbr de tâche, nr de tâche affichéees aprè filtrage, total RAF). la màj de la barre d'état était déclenchée qd on transforme le DTO en Bean. On contourne ce pbm en transformant le DTO en Bean (donc mettant à jour la barre d'état) après avoir fermé le ProgressDialog. NB : Reste que la liste des ressrouces humaines est tjs vide suite à l'import des tâches...

+++ planCharge-metier/src/main/java/fr/gouv/agriculture/dal/ct/metier/service/RapportService.java
@@ -10,10 +10,10 @@ public interface RapportService {
     String getAvancement();
 
 
-    void setProgressionMax(int progressionMax);
+    void setProgressionMax(long progressionMax);
 
-    int getProgressionMax();
+    long getProgressionMax();
 
 
-    void setProgressionCourante(int progressionCourante);
+    void setProgressionCourante(long progressionCourante);
 }
